requirements.txt
# Video Streaming Pipeline - Python Dependencies

# Web framework & server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Configuration management
pyyaml==6.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Async support
aiofiles==23.2.1
httpx==0.27.0

# Storage (optional S3 support)
boto3==1.34.131

# Video/media utilities
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
click==8.1.7

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
